Renaming ExceptionReporter file's to appropriate file names
